simplify graph inheritance
