+++ app/assets/javascripts/extensions/views/timeseries-graph/callout.js
@@ -31,17 +31,6 @@ function (Callout) {
       }
     },
 
-    weekHeader: function (el, group, groupIndex, model, index) {
-      var start = model.get('_start_at');
-      var end = moment(model.get('_end_at')).subtract(1, 'days');
-
-      return [
-        start.format(start.month() === end.month() ? 'D' : 'D MMM'),
-        ' to ',
-        end.format('D MMM YYYY')
-      ].join('');
-    },
-
     renderContent: function (el, group, groupIndex, model, index) {
       
       var header = $('<h3>').html(this.getHeader.apply(this, arguments));
