+++ app/assets/javascripts/extensions/views/timeseries-graph/multi-timeseries-graph.js
@@ -1,22 +1,30 @@
 define([
-  'require',
-  './timeseries-graph',
-  'extensions/views/graph/linelabel'
+  'extensions/views/graph/graph'
 ],
-function (require, TimeseriesGraph, LineLabel) {
-  var MultiTimeseriesGraph = TimeseriesGraph.extend({
+function (Graph) {
+  var MultiTimeseriesGraph = Graph.extend({
     
     components: function () {
       return [
         { view: this.sharedComponents.xaxis },
         { view: this.sharedComponents.yaxis },
-        { view: LineLabel },
-        { view: this.sharedComponents.line },
+        { view: this.sharedComponents.linelabel },
+        {
+          view: this.sharedComponents.line,
+          options: {
+            interactive: function (e) {
+              return e.slice % 3 !== 2;
+            }
+          }
+        },
         { view: this.sharedComponents.callout },
         { view: this.sharedComponents.hover }
       ];
-    }
+    },
     
+    getConfigNames: function () {
+      return ['overlay', this.collection.query.get('period') || 'week'];
+    }
   });
   
   return MultiTimeseriesGraph;
