--- app/assets/javascripts/extensions/views/timeseries-graph/stack.js
@@ -1,11 +1,0 @@
-define([
-  'extensions/views/graph/stack'
-],
-function (Stack) {
-  var TimeseriesStack = Stack.extend({
-    
-    interactive: true
-  });
-  
-  return TimeseriesStack;
-});
